MPDF auf Version 5.3 aktualisiert

+++ scrape/lib/mpdf50/compress.php
@@ -1,7 +1,7 @@
 <?php
 
 $excl = array( 'TABLES', 'LISTS', 'IMAGES-CORE', 
-'IMAGES-WMF', 'TABLES-ADVANCED-BORDERS', 'HTMLHEADERS-FOOTERS', 'COLUMNS', 'TOC', 'INDEX', 'BOOKMARKS', 'BARCODES', 'FORMS', 'WATERMARK', 'RTL', 'INDIC', 'ANNOTATIONS', 'GRADIENTS', 'BACKGROUND-IMAGES', 'CSS-FLOAT', 'CSS-IMAGE-FLOAT', 'CSS-POSITION', 'CSS-PAGE', 'BORDER-RADIUS', 'HYPHENATION', 'ENCRYPTION', 'DIRECTW', 'IMPORTS');
+'IMAGES-BMP', 'IMAGES-WMF', 'TABLES-ADVANCED-BORDERS', 'HTMLHEADERS-FOOTERS', 'COLUMNS', 'TOC', 'INDEX', 'BOOKMARKS', 'BARCODES', 'FORMS', 'WATERMARK', 'CJK-FONTS', 'RTL', 'INDIC', 'ANNOTATIONS', 'BACKGROUNDS', 'CSS-FLOAT', 'CSS-IMAGE-FLOAT', 'CSS-POSITION', 'CSS-PAGE', 'BORDER-RADIUS', 'HYPHENATION', 'ENCRYPTION', 'DIRECTW', 'IMPORTS', 'PROGRESS-BAR');
 
 
 	// *DIRECTW* = Write, WriteText, WriteCell, Text, Shaded_box, AutosizeText
@@ -18,7 +18,7 @@ if (!isset($_POST['generate']) || $_POST['generate']!='generate') {
 
 
 if (!file_exists('mpdf_source.php')) {
-	echo("ERROR - Could not find mpdf_source.php file in current directory. Please rename mpdf.php as mpdf_source.php"); 
+	die("ERROR - Could not find mpdf_source.php file in current directory. Please rename mpdf.php as mpdf_source.php"); 
 }
 
 
@@ -52,7 +52,7 @@ function checkedAll (frm1) {
 <ul>
 <li>For WMF Images, you must include both IMAGES-CORE and IMAGES-WMF</li>
 <li>JPG, PNG and JPG images are supported with IMAGES-CORE</li>
-<li>IMAGES-CORE are required for BACKGROUND-IMAGES or WATERMARKS to work</li>
+<li>IMAGES-CORE are required for BACKGROUNDS (IMAGES) or WATERMARKS to work</li>
 <li>DIRECTW includes the functions to Write directly to the PDF file e.g. Write, WriteText, WriteCell, Text, Shaded_box, AutosizeText</li>
 </ul>
 </div>
@@ -87,7 +87,7 @@ if (is_array($inc) && count($inc)>0 ) {
 set_magic_quotes_runtime(0);
 
 $l = file('mpdf_source.php');
-if (!count($l)) { echo("ERROR - Could not find mpdf_source.php file in current directory"); }
+if (!count($l)) { die("ERROR - Could not find mpdf_source.php file in current directory"); }
 $exclflags = array();
 $x = '';
 foreach($l AS $k=>$ln) {
@@ -120,9 +120,16 @@ foreach($l AS $k=>$ln) {
 		$x .= $ln; 
 	}
 }
-
-$check = file_put_contents('mpdf.php', $x);
-if (!$check) { echo("ERROR - Could not write to mpdf.php file. Are permissions correctly set?"); }
+// mPDF 5.0
+if (function_exists('file_put_contents')) {
+	$check = file_put_contents('mpdf.php', $x);
+}
+else {
+	$f=fopen('mpdf.php', 'w');
+	$check = fwrite($f, $x);
+	fclose($f);
+}
+if (!$check) { die("ERROR - Could not write to mpdf.php file. Are permissions correctly set?"); }
 echo '<p><b>mPDF file generated successfully!</b></p>';
 echo '<div>mPDF file size '.number_format((strlen($x)/1024)).' kB</div>';
